+++ vscode/CHANGELOG.md
@@ -5,6 +5,8 @@ file. Format inspired by https://keepachangelog.com/en/1.0.0/
 
 
 ## [Unreleased]
+
+## [0.0.12] - 2020-05-10
 ### Added
 - tag completion within notes